refactor(value): rename _shipped→_value, drop CHANGELOG/registry, keep label join
